+++ tests/testthat/test-plumber.R
@@ -3,11 +3,12 @@ context("Plumber")
 test_that("Endpoints are properly identified", {
   r <- plumber$new("files/endpoints.R")
   expect_equal(length(r$endpoints), 1)
-  expect_equal(length(r$endpoints[[1]]), 4)
+  expect_equal(length(r$endpoints[[1]]), 5)
   expect_equal(r$endpoints[[1]][[1]]$exec(), 5)
-  expect_equal(r$endpoints[[1]][[2]]$exec(), 10)
-  expect_equal(r$endpoints[[1]][[3]]$exec(), 12)
-  expect_equal(r$endpoints[[1]][[4]]$exec(), 14)
+  expect_equal(r$endpoints[[1]][[2]]$exec(), 5)
+  expect_equal(r$endpoints[[1]][[3]]$exec(), 10)
+  expect_equal(r$endpoints[[1]][[4]]$exec(), 12)
+  expect_equal(r$endpoints[[1]][[5]]$exec(), 14)
 })
 
 test_that("Empty file is OK", {
@@ -25,15 +26,16 @@ test_that("The file is sourced in the envir", {
 test_that("Verbs translate correctly", {
   r <- plumber$new("files/verbs.R")
   expect_equal(length(r$endpoints), 1)
-  expect_equal(length(r$endpoints[[1]]), 8)
+  expect_equal(length(r$endpoints[[1]]), 9)
   expect_equal(r$endpoints[[1]][[1]]$verbs, c("GET", "PUT", "POST", "DELETE", "HEAD", "OPTIONS"))
   expect_equal(r$endpoints[[1]][[2]]$verbs, "GET")
   expect_equal(r$endpoints[[1]][[3]]$verbs, "PUT")
   expect_equal(r$endpoints[[1]][[4]]$verbs, "POST")
   expect_equal(r$endpoints[[1]][[5]]$verbs, "DELETE")
-  expect_equal(r$endpoints[[1]][[6]]$verbs, c("POST", "GET"))
-  expect_equal(r$endpoints[[1]][[7]]$verbs, "HEAD")
-  expect_equal(r$endpoints[[1]][[8]]$verbs, "OPTIONS")
+  expect_equal(r$endpoints[[1]][[6]]$verbs, "POST")
+  expect_equal(r$endpoints[[1]][[7]]$verbs, "GET")
+  expect_equal(r$endpoints[[1]][[8]]$verbs, "HEAD")
+  expect_equal(r$endpoints[[1]][[9]]$verbs, "OPTIONS")
 })
 
 test_that("Invalid file fails gracefully", {
@@ -42,19 +44,19 @@ test_that("Invalid file fails gracefully", {
 
 test_that("plumb accepts a file", {
   r <- plumb("files/endpoints.R")
-  expect_length(r$endpoints[[1]], 4)
+  expect_length(r$endpoints[[1]], 5)
 })
 
 test_that("plumb accepts a directory with a `plumber.R` file", {
   # works without trailing slash
   r <- plumb(dir = 'files')
   expect_equal(length(r$endpoints), 1)
-  expect_equal(length(r$endpoints[[1]]), 4)
+  expect_equal(length(r$endpoints[[1]]), 5)
 
   # works with trailing slash
   r <- plumb(dir = 'files/')
   expect_equal(length(r$endpoints), 1)
-  expect_equal(length(r$endpoints[[1]]), 4)
+  expect_equal(length(r$endpoints[[1]]), 5)
 
   # errors when no plumber.R found
   expect_error(plumb(dir = 'files/static'), regexp="No plumber.R file found in the specified directory: files/static")
@@ -93,11 +95,12 @@ test_that("Empty endpoints error", {
 test_that("The old roxygen-style comments work", {
   r <- plumber$new("files/endpoints-old.R")
   expect_equal(length(r$endpoints), 1)
-  expect_equal(length(r$endpoints[[1]]), 4)
+  expect_equal(length(r$endpoints[[1]]), 5)
   expect_equal(r$endpoints[[1]][[1]]$exec(), 5)
-  expect_equal(r$endpoints[[1]][[2]]$exec(), 10)
-  expect_equal(r$endpoints[[1]][[3]]$exec(), 12)
-  expect_equal(r$endpoints[[1]][[4]]$exec(), 14)
+  expect_equal(r$endpoints[[1]][[2]]$exec(), 5)
+  expect_equal(r$endpoints[[1]][[3]]$exec(), 10)
+  expect_equal(r$endpoints[[1]][[4]]$exec(), 12)
+  expect_equal(r$endpoints[[1]][[5]]$exec(), 14)
 })
 
 test_that("routes can be constructed correctly", {